requirements.txt
requests==2.23.0
python-telegram-bot==13.4.1
pymongo==3.11.3